Including someone's email like that is a no-no
Including an email where some random script can scrape it and start spamming it is a big faux pas. Usually some amount of obfuscation is in order, but even those are easily defeated. More importantly, the message explicitly asks for discussion to happen at the link provided, so questions emphatically *should not be directed to my email*